+++ plugins/main/public/controllers/management/components/management/configuration/configuration-settings.js
@@ -11,7 +11,7 @@
  */
 
 import { hasAgentSupportModule } from '../../../../../react-services/wz-agents';
-import { WAZUH_MODULES_ID } from '../../../../../../common/constants'
+import { WAZUH_MODULES_ID } from '../../../../../../common/constants';
 
 export default [
   {
@@ -22,52 +22,52 @@ export default [
         name: 'Global Configuration',
         description: 'Global and remote settings',
         goto: 'global-configuration',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Cluster',
         description: 'Master node configuration',
         goto: 'cluster',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Registration Service',
         description: 'Automatic agent registration service',
         goto: 'registration-service',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Global Configuration',
         description: 'Logging settings that apply to the agent',
         goto: 'global-configuration-agent',
-        when: 'agent'
+        when: 'agent',
       },
       {
         name: 'Communication',
         description: 'Settings related to the connection with the manager',
         goto: 'client',
-        when: 'agent'
+        when: 'agent',
       },
       {
         name: 'Anti-flooding settings',
         description: 'Agent bucket parameters to avoid event flooding',
         goto: 'client-buffer',
-        when: 'agent'
+        when: 'agent',
       },
       {
         name: 'Labels',
         description:
           'User-defined information about the agent included in alerts',
         goto: 'alerts-agent',
-        when: 'agent'
-      }
+        when: 'agent',
+      },
       // ,
       // { //TODO: Uncomment this to activate Log Settings
       //   name: 'Log settings',
       //   description: 'Alerts, archives and internal settings',
       //   goto: 'log-settings'
       // }
-    ]
+    ],
   },
   {
     title: 'Alerts and output management',
@@ -77,16 +77,16 @@ export default [
         name: 'Alerts',
         description: 'Settings related to the alerts and their format',
         goto: 'alerts',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Integrations',
         description:
           'Slack, VirusTotal and PagerDuty integrations with external APIs',
         goto: 'integrations',
-        when: 'manager'
-      }
-    ]
+        when: 'manager',
+      },
+    ],
   },
   {
     title: 'Auditing and policy monitoring',
@@ -96,22 +96,22 @@ export default [
         name: 'Policy monitoring',
         description:
           'Configuration to ensure compliance with security policies, standards and hardening guides',
-        goto: 'policy-monitoring'
+        goto: 'policy-monitoring',
       },
       {
         name: 'OpenSCAP',
         description:
           'Configuration assessment and automation of compliance monitoring using SCAP checks',
         goto: 'open-scap',
-        when: agent => hasAgentSupportModule(agent, WAZUH_MODULES_ID.OPEN_SCAP)
+        when: agent => hasAgentSupportModule(agent, WAZUH_MODULES_ID.OPEN_SCAP),
       },
       {
         name: 'CIS-CAT',
         description:
           'Configuration assessment using CIS scanner and SCAP checks',
-        goto: 'cis-cat'
-      }
-    ]
+        goto: 'cis-cat',
+      },
+    ],
   },
   {
     title: 'System threats and incident response',
@@ -122,45 +122,45 @@ export default [
         description:
           'Discover what applications are affected by well-known vulnerabilities',
         goto: 'vulnerabilities',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Osquery',
         description:
           'Expose an operating system as a high-performance relational database',
-        goto: 'osquery'
+        goto: 'osquery',
       },
       {
         name: 'Inventory data',
         description:
           'Gather relevant information about system operating system, hardware, networking and packages',
-        goto: 'inventory'
+        goto: 'inventory',
       },
       {
-        name: 'Active Response',
+        name: 'Active response',
         description: 'Active threat addressing by immediate response',
         goto: 'active-response',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Active response',
         description: 'Active threat addressing by immediate response',
         goto: 'active-response-agent',
-        when: 'agent'
+        when: 'agent',
       },
       {
         name: 'Commands',
         description: 'Configuration options of the Command wodle',
-        goto: 'commands'
+        goto: 'commands',
       },
       {
         name: 'Docker listener',
         description:
           'Monitor and collect the activity from Docker containers such as creation, running, starting, stopping or pausing events',
         goto: 'docker-listener',
-        when: agent => hasAgentSupportModule(agent, WAZUH_MODULES_ID.DOCKER)
-      }
-    ]
+        when: agent => hasAgentSupportModule(agent, WAZUH_MODULES_ID.DOCKER),
+      },
+    ],
   },
   {
     title: 'Log data analysis',
@@ -170,22 +170,22 @@ export default [
         name: 'Log collection',
         description:
           'Log analysis from text files, Windows events or syslog outputs',
-        goto: 'log-collection'
+        goto: 'log-collection',
       },
       {
         name: 'Integrity monitoring',
         description:
           'Identify changes in content, permissions, ownership, and attributes of files',
-        goto: 'integrity-monitoring'
+        goto: 'integrity-monitoring',
       },
       {
         name: 'Agentless',
         description:
           'Run integrity checks on devices such as routers, firewalls and switches',
         goto: 'agentless',
-        when: 'manager'
-      }
-    ]
+        when: 'manager',
+      },
+    ],
   },
   {
     title: 'Cloud security monitoring',
@@ -195,32 +195,30 @@ export default [
         name: 'Amazon S3',
         description:
           'Security events related to Amazon AWS services, collected directly via AWS API',
-        goto: 'aws-s3'
+        goto: 'aws-s3',
       },
       {
         name: 'Azure Logs',
         description: 'Configuration options of the Azure Logs wodle',
         goto: 'azure-logs',
-        when: 'manager'
+        when: 'manager',
       },
       {
         name: 'Google Cloud Pub/Sub',
         description: 'Configuration options of the Google Cloud Pub/Sub module',
-        goto: 'gcp-pubsub'
+        goto: 'gcp-pubsub',
       },
       {
         name: 'GitHub',
-        description:
-          'Detect threats targeting GitHub organizations',
-        goto: 'github'
+        description: 'Detect threats targeting GitHub organizations',
+        goto: 'github',
       },
       {
         name: 'Office 365',
-        description:
-          'Configuration options of the Office 365 module',
+        description: 'Configuration options of the Office 365 module',
         goto: 'office365',
-        when: 'manager'
-      }
-    ]
-  }
+        when: 'manager',
+      },
+    ],
+  },
 ];